pyproject.toml
[project]
name = "telegram-bot-template"
version = "2.4.0"
description = "Best way to create a scalable telegram bot with analytics"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "donBarbos", email = "donbarbos@proton.me"},
]
package-mode = false
requires-python = "<4.0,>=3.10"
dependencies = [
    "sqlalchemy<3.0.0,>=2.0.38",
    "pydantic<3.0.0,>=2.10.6",
    "pydantic-settings<3.0.0,>=2.8.1",
    "pgvector (>=0.4.1,<0.5.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
]

[project.urls]
homepage = "https://github.com/donBarbos/telegram-bot-template"
repository = "https://github.com/donBarbos/telegram-bot-template"

[dependency-groups]
bot = [
    "aiogram>=3.19.0,<4.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "redis>=5.2.1,<8.0.0",
    "loguru>=0.7.3,<1.0.0",
    "sentry-sdk[loguru]>=2.20.0,<3.0.0",
    "aiohttp[speedups]>=3.11.16,<4.0.0",
    "babel>=2.17.0,<3.0.0",
    "cachetools>=5.5.1,<7.0.0",
    "orjson>=3.10.15,<4.0.0",
    "alembic>=1.14.1,<2.0.0",
    "prometheus-client>=0.21.1,<1.0.0",
    "pydantic-ai>=1.0.0,<2.0.0",
    "pgvector>=0.2.4,<1.0.0",
]
admin = [
    "gunicorn>=23.0.0,<24.0.0",
    "flask>=3.1.0,<4.0.0",
    "flask-admin>=1.6.1,<3.0.0",
    "flask-security-too>=5.5.2,<6.0.0",
    "flask-caching>=2.3.0,<3.0.0",
    "flask-babel>=4.0.0,<5.0.0",
    "flask-sqlalchemy>=3.1.1,<4.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "tablib[xlsx]>=3.8.0,<4.0.0",
]
dev = [
    "ruff>=0.9.5,<1.0.0",
    "mypy>=1.15.0,<2.0.0",
    "pre-commit>=4.2.0,<5.0.0",
    "types-cachetools>=5.5.0.20240820,<7.0.0.0",
]

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 120
lint.select = ["ALL"]
lint.ignore = ["D", "ANN401", "FIX002", "COM812", "ISC001", "FBT001", "FBT002", "ERA", "ARG005", "PGH003", "A005"]
cache-dir = "/tmp/ruff-cache/"
extend-exclude = ["logs", "configs", "migrations"]

[tool.ruff.lint.isort]
no-lines-before = ["future", "standard-library"]
known-local-folder = ["bot", "admin"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/*.py" = ["ANN401", "S101", "S311"]

[tool.mypy]
python_version = "3.10"
files = "bot/*.py"
plugins = ["pydantic.mypy"]
ignore_missing_imports = false
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
show_error_context = true
pretty = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_imports_for_stubs = true
namespace_packages = true
show_absolute_path = true

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true
